Additional testing with mocked http via pook added
