+++ pyslackers_website/slack/tests/conftest.py
@@ -1,0 +1,1 @@
+import pytest
