+++ requirements.txt
@@ -22,6 +22,7 @@ Markdown==2.6.9
 mccabe==0.6.1
 msgpack-python==0.4.8
 oauthlib==2.0.3
+pook==0.2.3
 psycopg2==2.7.3.1
 py==1.4.34
 pycodestyle==2.3.1
